Boss attack fix

+++ Scripts/skeleton_boss.gd
@@ -8,7 +8,7 @@ var freezing = false
 @onready var animation_player = $AnimationPlayer
 
 func _physics_process(_delta):
-	
+	print("walk: ",finished_walk, " progress: ",get_parent_node_3d().progress, " max: ",get_parent_node_3d().curve_length, " hp: ",health)
 	if finished_walk:
 		play_attack_animation()
 		if can_attack:
@@ -34,11 +34,11 @@ func change_speed(value:float)->void:
 #Function that happens when mob is struck with arrow/spell/canon
 func take_damage(dmg: int) -> void:
 	health-=dmg
-	if(health<=0):
+	if (health<=0) and (health+dmg) > 0:
 		get_parent_node_3d().delete_object()
 
 #Resetting attacking 
 func _on_attack_cd_timeout() -> void:
-	take_damage(100)
+	take_damage(200)
 	get_parent_node_3d().get_parent_node_3d().get_parent_node_3d().take_damage(damage)
 	can_attack=true
